+++ x/gov/keeper/keeper.go
@@ -183,3 +183,12 @@ func (k Keeper) assertMetadataLength(metadata string) error {
 	}
 	return nil
 }
+
+// assertSummaryLength returns an error if given summary length
+// is greater than a pre-defined 40*MaxMetadataLen.
+func (keeper Keeper) assertSummaryLength(summary string) error {
+	if summary != "" && uint64(len(summary)) > 40*keeper.config.MaxMetadataLen {
+		return types.ErrSummaryTooLong.Wrapf("got summary with length %d", len(summary))
+	}
+	return nil
+}