+++ x/group/keeper/msg_server.go
@@ -521,7 +521,7 @@ func (k Keeper) SubmitProposal(goCtx context.Context, msg *group.MsgSubmitPropos
 		return nil, err
 	}
 
-	if err := k.assertMetadataLength(msg.Summary, "proposal summary"); err != nil {
+	if err := k.assertSummaryLength(msg.Summary); err != nil {
 		return nil, err
 	}
 
@@ -1058,6 +1058,15 @@ func (k Keeper) doUpdateGroup(ctx sdk.Context, groupID uint64, reqGroupAdmin str
 func (k Keeper) assertMetadataLength(metadata, description string) error {
 	if metadata != "" && uint64(len(metadata)) > k.config.MaxMetadataLen {
 		return errorsmod.Wrapf(errors.ErrMaxLimit, description)
+	}
+	return nil
+}
+
+// assertSummaryLength returns an error if given summary length
+// is greater than a pre-defined 40*MaxMetadataLen.
+func (k Keeper) assertSummaryLength(summary string) error {
+	if summary != "" && uint64(len(summary)) > 40*k.config.MaxMetadataLen {
+		return errorsmod.Wrapf(errors.ErrMaxLimit, "proposal summary is too long")
 	}
 	return nil
 }